requirements.txt
streamlit==1.43.2
pandas==2.2.3
matplotlib==3.10.1
pillow==11.1.0